requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.1.0
click==8.0.3
comm==0.2.1
common==0.1.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
executing==2.0.1
fastapi==0.115.6
fastjsonschema==2.19.1
filelock==3.15.1
Flask==2.0.2
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gguf==0.6.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.23.4
idna==3.4
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
load-dotenv==0.1.0
marko==2.0.2
MarkupSafe==2.0.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
mutagen==1.47.0
mysql-connector==2.2.9
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.24.4
openai==1.10.0
opencv-python==4.10.0.82
overrides==7.6.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.3
scipy==1.13.1
seaborn==0.13.1
Send2Trash==1.8.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.41.3
sympy==1.12.1
terminado==0.18.0
tinycss2==1.2.1
tokenizers==0.19.1
torch==2.1.2
torchvision==0.18.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.41.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.2.34
ultralytics-thop==2.0.0
uri-template==1.3.0
urllib3==2.0.2
uvicorn==0.32.1
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==13.1
Werkzeug==2.0.2
widgetsnbextension==4.0.9
yarl==1.9.4
yt-dlp==2024.10.22
